added modals for timer, example list of washing machines, added washingmachine type to RootStackParamList, improved styling
