+++ types/types.ts
@@ -7,7 +7,6 @@ export type RootStackParamList = {
     Settings: undefined;     // No parameters for Settings screen
     ListIssues: undefined;   // No parameters for ListIssues screen
     WashingMachine: undefined; // No parameters for WashingMachine screen
-    ManageWashingMachine: undefined; // No parameters for CreateWashingMachine screen
   };
 
 export type AuthStackParamList = {
